requirements.txt
# Core dependencies
numpy>=1.19.0
open3d>=0.15.0
matplotlib>=3.3.0
pillow>=8.0.0
tqdm>=4.50.0

# 3D model processing
trimesh>=3.9.0
pyglet>=1.5.0

# Image analysis and OCR
opencv-python>=4.5.0
pytesseract>=0.3.8
requests>=2.25.0

# Vector database and embeddings
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# Machine learning
scikit-learn>=0.24.0

# Utilities
pathlib>=1.0.1

# Additional dependencies
protobuf==3.20.0
tensorflow>=2.5.0
regex>=2021.4.4 